requirements.txt
py-cord==2.0.0b4
Pillow
requests
apscheduler